In this commit I added the architecture to be used in the test. I decided to use an architecture by features so that in this way the project is more scalable and easier to understand for other developers.
Also I added some of the libraries that I will use for the user interface, at the moment I have not added any library to track the user's face and the camera. The previous libraries will be added in the next commits.

I added some unit tests to the network and data layers only to guarantee the quality of the code in these two layers